test(lib-versioning): split the analyze-commits spec into focused suites
