+++ libs/versioning/src/flow/steps/analyze-commits.spec.ts
@@ -1,148 +1,7 @@
-import type { Logger } from '@hyperfrontend/logging'
-import type { Tree } from '@hyperfrontend/project-scope/vfs'
-
 import type { GitClient } from '../../git/factory'
-import type { Registry } from '../../registry/models/registry'
-import type { FlowConfig, FlowContext, FlowState } from '../models/types'
-
+import type { FlowContext } from '../models/types'
+import { createMockContext, createMockGitClient, createMockLogger } from './__test-utils__/analyze-commits-mocks'
 import { createAnalyzeCommitsStep, ANALYZE_COMMITS_STEP_ID } from './analyze-commits'
-
-function createMockLogger(): Logger {
-  return {
-    info: jest.fn(),
-    warn: jest.fn(),
-    error: jest.fn(),
-    debug: jest.fn(),
-    setLogLevel: jest.fn(),
-  } as unknown as Logger
-}
-
-function createMockTree(): Tree {
-  return {
-    root: '/workspace',
-    read: () => null,
-    write: jest.fn(),
-    exists: () => false,
-    delete: jest.fn(),
-    rename: jest.fn(),
-    isFile: () => false,
-    children: () => [],
-    listChanges: () => [],
-  } as unknown as Tree
-}
-
-function createMockRegistry(): Registry {
-  return {
-    name: 'mock',
-    url: 'https://mock.registry.com',
-    async getLatestVersion() {
-      return null
-    },
-    async isVersionPublished() {
-      return false
-    },
-    async getPackageInfo() {
-      return null
-    },
-    async getVersionInfo() {
-      return null
-    },
-    async listVersions() {
-      return []
-    },
-  }
-}
-
-interface MockGitClientOptions {
-  commits?: readonly { message: string; hash: string }[]
-  packageTags?: readonly { name: string; hash: string }[]
-  projectTags?: readonly { name: string; hash: string }[]
-  commitReachable?: boolean
-}
-
-function createMockGitClient(options: MockGitClientOptions = {}): GitClient {
-  const { commits = [], packageTags = [], projectTags = [], commitReachable = true } = options
-
-  return {
-    cwd: '/workspace',
-    timeout: 30000,
-    getCommitLog: (opts?: { maxCount?: number }) => {
-      if (opts?.maxCount) {
-        return commits.slice(0, opts.maxCount)
-      }
-      return commits
-    },
-    getCommitsBetween: () => commits,
-    getCommitsSince: () => commits,
-    getCommit: () => (commits.length > 0 ? commits[0] : null),
-    commitExists: () => true,
-    commitReachableFromHead: () => commitReachable,
-    getTags: () => [],
-    getTag: () => null,
-    createTag: jest.fn(),
-    deleteTag: () => true,
-    tagExists: () => false,
-    getLatestTag: () => null,
-    getTagsForPackage: (name: string) => {
-      if (name.startsWith('@') || name.includes('/')) {
-        return packageTags as { name: string; hash: string }[]
-      }
-      return projectTags as { name: string; hash: string }[]
-    },
-    pushTag: () => true,
-    createCommit: jest.fn(),
-    stage: () => true,
-    unstage: () => true,
-    stageAll: () => true,
-    amendCommit: jest.fn(),
-    createEmptyCommit: jest.fn(),
-    getHead: () => 'abc123',
-    getCurrentBranch: () => 'main',
-    hasStagedChanges: () => false,
-    hasUnstagedChanges: () => false,
-    hasUntrackedFiles: () => false,
-    getStatus: () => ({
-      clean: true,
-      entries: [],
-      staged: [],
-      unstaged: [],
-      untracked: [],
-    }),
-    isClean: () => true,
-    getRepositoryRoot: () => '/workspace',
-    getHeadHash: () => 'abc123',
-    getHeadShortHash: () => 'abc123',
-    getModifiedFiles: () => [],
-    getUntrackedFiles: () => [],
-    getStagedFiles: () => [],
-  } as unknown as GitClient
-}
-
-function createMockContext(overrides?: {
-  git?: GitClient
-  logger?: Logger
-  state?: FlowState
-  config?: Partial<FlowConfig>
-  projectName?: string
-  packageName?: string
-}): FlowContext {
-  return {
-    workspaceRoot: '/workspace',
-    projectName: overrides?.projectName ?? 'lib-test',
-    projectRoot: '/workspace/libs/test',
-    packageName: overrides?.packageName ?? '@test/pkg',
-    tree: createMockTree(),
-    registry: createMockRegistry(),
-    git: overrides?.git ?? createMockGitClient(),
-    logger: overrides?.logger ?? createMockLogger(),
-    config: {
-      preset: 'conventional',
-      releaseTypes: ['feat', 'fix', 'perf', 'revert'],
-      ...overrides?.config,
-    },
-    state: overrides?.state ?? {},
-  }
-}
 
 describe('analyze-commits step', () => {
   describe('ANALYZE_COMMITS_STEP_ID', () => {
@@ -460,283 +319,6 @@ describe('analyze-commits step', () => {
         })
       })
 
-      describe('scope filtering strategies', () => {
-        it('uses hybrid strategy by default', async () => {
-          const git = createMockGitClient({
-            commits: [{ message: 'feat: new feature', hash: 'commit1' }],
-          })
-          const context = createMockContext({
-            git,
-            state: { isFirstRelease: true },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          const result = await step.execute(context)
-
-          expect(result.status).toBe('success')
-          expect(result.message).toContain('strategy: hybrid')
-        })
-
-        it('respects scope-only strategy from config', async () => {
-          const git = createMockGitClient({
-            commits: [
-              { message: 'feat(lib-test): scoped feature', hash: 'commit1' },
-              { message: 'feat: unscoped feature', hash: 'commit2' },
-            ],
-          })
-          const context = createMockContext({
-            git,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: { strategy: 'scope-only' },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          const result = await step.execute(context)
-
-          expect(result.status).toBe('success')
-          expect(result.message).toContain('strategy: scope-only')
-          expect(result.stateUpdates?.commits).toHaveLength(1)
-        })
-
-        it('respects file-only strategy from config', async () => {
-          const git = createMockGitClient({
-            commits: [
-              { message: 'feat(lib-test): scoped feature', hash: 'commit1' },
-              { message: 'feat: unscoped feature', hash: 'commit2' },
-            ],
-          })
-          const context = createMockContext({
-            git,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: { strategy: 'file-only' },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          const result = await step.execute(context)
-
-          expect(result.status).toBe('success')
-          expect(result.message).toContain('strategy: file-only')
-        })
-
-        it('infers scope-only strategy when >70% commits have scopes', async () => {
-          const commits = [
-            { message: 'feat(scope1): feature 1', hash: 'c1' },
-            { message: 'feat(scope2): feature 2', hash: 'c2' },
-            { message: 'fix(scope3): fix 1', hash: 'c3' },
-            { message: 'feat(scope4): feature 3', hash: 'c4' },
-            { message: 'fix(scope5): fix 2', hash: 'c5' },
-            { message: 'feat(scope6): feature 4', hash: 'c6' },
-            { message: 'fix(scope7): fix 3', hash: 'c7' },
-            { message: 'feat(scope8): feature 5', hash: 'c8' },
-            { message: 'feat: unscoped 1', hash: 'c9' },
-            { message: 'fix: unscoped 2', hash: 'c10' },
-          ]
-          const git = createMockGitClient({ commits })
-          const context = createMockContext({
-            git,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: { strategy: 'inferred' },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          const result = await step.execute(context)
-
-          expect(result.status).toBe('success')
-          expect(result.message).toContain('strategy: scope-only')
-        })
-
-        it('infers file-only strategy when <30% commits have scopes', async () => {
-          const commits = [
-            { message: 'feat(scope1): feature 1', hash: 'c1' },
-            { message: 'feat(scope2): feature 2', hash: 'c2' },
-            { message: 'fix: fix 1', hash: 'c3' },
-            { message: 'feat: feature 3', hash: 'c4' },
-            { message: 'fix: fix 2', hash: 'c5' },
-            { message: 'feat: feature 4', hash: 'c6' },
-            { message: 'fix: fix 3', hash: 'c7' },
-            { message: 'feat: feature 5', hash: 'c8' },
-            { message: 'feat: unscoped 1', hash: 'c9' },
-            { message: 'fix: unscoped 2', hash: 'c10' },
-          ]
-          const git = createMockGitClient({ commits })
-          const context = createMockContext({
-            git,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: { strategy: 'inferred' },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          const result = await step.execute(context)
-
-          expect(result.status).toBe('success')
-          expect(result.message).toContain('strategy: file-only')
-        })
-
-        it('infers hybrid strategy when scope ratio is between 30-70%', async () => {
-          const commits = [
-            { message: 'feat(scope1): feature 1', hash: 'c1' },
-            { message: 'feat(scope2): feature 2', hash: 'c2' },
-            { message: 'fix(scope3): fix 1', hash: 'c3' },
-            { message: 'feat(scope4): feature 3', hash: 'c4' },
-            { message: 'fix(scope5): fix 2', hash: 'c5' },
-            { message: 'feat: feature 4', hash: 'c6' },
-            { message: 'fix: fix 3', hash: 'c7' },
-            { message: 'feat: feature 5', hash: 'c8' },
-            { message: 'feat: unscoped 1', hash: 'c9' },
-            { message: 'fix: unscoped 2', hash: 'c10' },
-          ]
-          const git = createMockGitClient({ commits })
-          const context = createMockContext({
-            git,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: { strategy: 'inferred' },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          const result = await step.execute(context)
-
-          expect(result.status).toBe('success')
-          expect(result.message).toContain('strategy: hybrid')
-        })
-
-        it('infers file-only strategy when no commits exist (ratio is 0)', async () => {
-          const git = createMockGitClient({ commits: [] })
-          const context = createMockContext({
-            git,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: { strategy: 'inferred' },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          const result = await step.execute(context)
-
-          expect(result.status).toBe('success')
-          expect(result.message).toContain('strategy: file-only')
-        })
-      })
-
-      describe('classification result', () => {
-        it('includes classificationResult in state updates', async () => {
-          const git = createMockGitClient({
-            commits: [
-              { message: 'feat(lib-test): scoped feature', hash: 'commit1' },
-              { message: 'feat: unscoped feature', hash: 'commit2' },
-            ],
-          })
-          const context = createMockContext({
-            git,
-            state: { isFirstRelease: true },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          const result = await step.execute(context)
-
-          expect(result.status).toBe('success')
-          expect(result.stateUpdates?.classificationResult).toBeDefined()
-          expect(result.stateUpdates?.classificationResult?.summary).toBeDefined()
-          expect(result.stateUpdates?.classificationResult?.summary.bySource).toBeDefined()
-        })
-
-        it('provides classification summary breakdown', async () => {
-          const git = createMockGitClient({
-            commits: [{ message: 'feat(lib-test): scoped feature', hash: 'commit1' }],
-          })
-          const logger = createMockLogger()
-          const context = createMockContext({
-            git,
-            logger,
-            state: { isFirstRelease: true },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          await step.execute(context)
-
-          expect(logger.debug).toHaveBeenCalledWith(expect.stringContaining('Classification breakdown:'))
-        })
-      })
-
-      describe('scope filtering config', () => {
-        it('applies excludeScopes from config', async () => {
-          const git = createMockGitClient({
-            commits: [
-              { message: 'feat(lib-test): included', hash: 'commit1' },
-              { message: 'feat(deps): excluded dep update', hash: 'commit2' },
-            ],
-          })
-          const context = createMockContext({
-            git,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: {
-                strategy: 'hybrid',
-                excludeScopes: ['deps'],
-              },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          const result = await step.execute(context)
-
-          expect(result.status).toBe('success')
-          const commits = result.stateUpdates?.commits ?? []
-          expect(commits.every((c) => !c.scope.includes('deps'))).toBe(true)
-        })
-
-        it('applies includeScopes from config', async () => {
-          const git = createMockGitClient({
-            commits: [
-              { message: 'feat(custom-scope): custom scoped', hash: 'commit1' },
-              { message: 'feat(other): other scope', hash: 'commit2' },
-            ],
-          })
-          const context = createMockContext({
-            git,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: {
-                strategy: 'scope-only',
-                includeScopes: ['custom-scope'],
-              },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          const result = await step.execute(context)
-
-          expect(result.status).toBe('success')
-          expect(result.stateUpdates?.commits).toHaveLength(1)
-        })
-
-        it('logs project scopes derivation', async () => {
-          const git = createMockGitClient({
-            commits: [{ message: 'feat: feature', hash: 'commit1' }],
-          })
-          const logger = createMockLogger()
-          const context = createMockContext({
-            git,
-            logger,
-            state: { isFirstRelease: true },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          await step.execute(context)
-
-          expect(logger.debug).toHaveBeenCalledWith(expect.stringContaining('Project scopes:'))
-        })
-      })
-
       describe('file-based commit detection', () => {
         it('detects commits touching project files in hybrid mode', async () => {
           const git = createMockGitClient({
@@ -853,178 +435,6 @@ describe('analyze-commits step', () => {
           const result = await step.execute(context)
 
           expect(result.message).toContain('3 total')
-        })
-      })
-
-      describe('infrastructure path-based detection', () => {
-        it('detects commits touching infrastructure paths', async () => {
-          const baseCommits = [
-            { message: 'feat(ci): update pipeline', hash: 'infra-commit1' },
-            { message: 'feat(lib-test): regular feature', hash: 'commit2' },
-          ]
-          const git = {
-            ...createMockGitClient({ commits: baseCommits }),
-            getCommitLog: (opts?: { maxCount?: number; path?: string }) => {
-              if (opts?.path === 'tools/') {
-                return [{ message: 'feat(ci): update pipeline', hash: 'infra-commit1' }]
-              }
-              if (opts?.maxCount) {
-                return baseCommits.slice(0, opts.maxCount)
-              }
-              return baseCommits
-            },
-            getCommitsSince: (tag: string, opts?: { path?: string }) => {
-              if (opts?.path === 'tools/') {
-                return [{ message: 'feat(ci): update pipeline', hash: 'infra-commit1' }]
-              }
-              return baseCommits
-            },
-          } as unknown as GitClient
-          const logger = createMockLogger()
-          const context = createMockContext({
-            git,
-            logger,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: {
-                infrastructure: {
-                  paths: ['tools/'],
-                },
-              },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          await step.execute(context)
-
-          expect(logger.debug).toHaveBeenCalledWith(expect.stringContaining('infrastructure paths'))
-        })
-
-        it('detects commits from multiple infrastructure paths', async () => {
-          const baseCommits = [
-            { message: 'feat(ci): update ci', hash: 'ci-commit' },
-            { message: 'feat(scripts): add script', hash: 'scripts-commit' },
-            { message: 'feat(lib-test): feature', hash: 'feature-commit' },
-          ]
-          const git = {
-            ...createMockGitClient({ commits: baseCommits }),
-            getCommitLog: (opts?: { maxCount?: number; path?: string }) => {
-              if (opts?.path === '.github/') {
-                return [{ message: 'feat(ci): update ci', hash: 'ci-commit' }]
-              }
-              if (opts?.path === 'scripts/') {
-                return [{ message: 'feat(scripts): add script', hash: 'scripts-commit' }]
-              }
-              if (opts?.maxCount) {
-                return baseCommits.slice(0, opts.maxCount)
-              }
-              return baseCommits
-            },
-          } as unknown as GitClient
-          const logger = createMockLogger()
-          const context = createMockContext({
-            git,
-            logger,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: {
-                infrastructure: {
-                  paths: ['.github/', 'scripts/'],
-                },
-              },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          await step.execute(context)
-
-          expect(logger.debug).toHaveBeenCalledWith(expect.stringContaining('infrastructure paths'))
-        })
-      })
-
-      describe('infrastructure scope-based detection', () => {
-        it('detects commits matching infrastructure scopes', async () => {
-          const commits = [
-            { message: 'feat(ci): update pipeline', hash: 'ci-commit' },
-            { message: 'feat(deps): update deps', hash: 'deps-commit' },
-            { message: 'feat(lib-test): feature', hash: 'feature-commit' },
-          ]
-          const git = createMockGitClient({ commits })
-          const logger = createMockLogger()
-          const context = createMockContext({
-            git,
-            logger,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: {
-                infrastructure: {
-                  scopes: ['ci', 'deps'],
-                },
-              },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          await step.execute(context)
-
-          expect(logger.debug).toHaveBeenCalledWith(expect.stringContaining('Infrastructure matcher'))
-        })
-      })
-
-      describe('infrastructure custom matcher', () => {
-        it('applies custom infrastructureMatcher function', async () => {
-          const commits = [
-            { message: 'feat(tooling): add tool', hash: 'tooling-commit' },
-            { message: 'feat(lib-test): feature', hash: 'feature-commit' },
-          ]
-          const git = createMockGitClient({ commits })
-          const logger = createMockLogger()
-          const customMatcher = jest.fn((ctx: { scope: readonly string[] }) => ctx.scope.includes('tooling'))
-          const context = createMockContext({
-            git,
-            logger,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: {
-                infrastructureMatcher: customMatcher,
-              },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          await step.execute(context)
-
-          expect(customMatcher).toHaveBeenCalled()
-          expect(logger.debug).toHaveBeenCalledWith(expect.stringContaining('Infrastructure matcher'))
-        })
-
-        it('combines infrastructure config matcher with custom matcher using OR logic', async () => {
-          const commits = [
-            { message: 'feat(ci): update ci', hash: 'ci-commit' },
-            { message: 'feat(custom): custom match', hash: 'custom-commit' },
-            { message: 'feat(lib-test): feature', hash: 'feature-commit' },
-          ]
-          const git = createMockGitClient({ commits })
-          const logger = createMockLogger()
-          const customMatcher = jest.fn((ctx: { scope: readonly string[] }) => ctx.scope.includes('custom'))
-          const context = createMockContext({
-            git,
-            logger,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: {
-                infrastructure: {
-                  scopes: ['ci'],
-                },
-                infrastructureMatcher: customMatcher,
-              },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          await step.execute(context)
-
-          expect(logger.debug).toHaveBeenCalledWith(expect.stringContaining('Infrastructure matcher'))
         })
       })
 
@@ -1179,162 +589,6 @@ describe('analyze-commits step', () => {
           const calls = getCommitsSinceMock.mock.calls as unknown as CommitsSinceCall[]
           const pathFilteredCalls = calls.filter((call) => call[1]?.path)
           expect(pathFilteredCalls.length).toBeGreaterThan(0)
-        })
-      })
-
-      describe('infrastructure returns undefined when not configured', () => {
-        it('returns undefined for infrastructureCommitHashes when no infrastructure config', async () => {
-          const git = createMockGitClient({
-            commits: [{ message: 'feat: feature', hash: 'commit1' }],
-          })
-          const context = createMockContext({
-            git,
-            state: { isFirstRelease: true },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          const result = await step.execute(context)
-
-          expect(result.status).toBe('success')
-          const logger = context.logger as jest.Mocked<Logger>
-          const debugCalls = logger.debug.mock.calls.map((call) => call[0])
-          const infraLogs = debugCalls.filter((msg: string) => msg.includes('infrastructure') || msg.includes('Infrastructure'))
-          expect(infraLogs).toHaveLength(0)
-        })
-      })
-
-      describe('infrastructure path detection with existing release', () => {
-        it('uses getCommitsSince for infrastructure paths when publishedCommit exists', async () => {
-          const baseCommits = [
-            { message: 'feat(ci): update ci', hash: 'ci-commit' },
-            { message: 'feat(lib-test): feature', hash: 'feature-commit' },
-          ]
-          const getCommitsSinceMock = jest.fn((commit: string, opts?: { path?: string }) => {
-            if (opts?.path === 'tools/') {
-              return [{ message: 'feat(ci): update ci', hash: 'ci-commit' }]
-            }
-            return baseCommits
-          })
-          const git = {
-            ...createMockGitClient({
-              commits: baseCommits,
-              commitReachable: true,
-            }),
-            getCommitsSince: getCommitsSinceMock,
-          } as unknown as GitClient
-          const logger = createMockLogger()
-          const context = createMockContext({
-            git,
-            logger,
-            state: {
-              isFirstRelease: false,
-              publishedCommit: 'abc1234',
-            },
-            config: {
-              scopeFiltering: {
-                infrastructure: {
-                  paths: ['tools/'],
-                },
-              },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          await step.execute(context)
-
-          type CommitsSinceCall = [commit: string, opts?: { path?: string }]
-          const calls = getCommitsSinceMock.mock.calls as unknown as CommitsSinceCall[]
-          const infraPathCalls = calls.filter((call) => call[1]?.path === 'tools/')
-          expect(infraPathCalls.length).toBeGreaterThan(0)
-          expect(logger.debug).toHaveBeenCalledWith(expect.stringContaining('infrastructure paths'))
-        })
-      })
-
-      describe('infrastructure path with no matching commits', () => {
-        it('logs zero commits when infrastructure paths configured but no matches', async () => {
-          const baseCommits = [{ message: 'feat(lib-test): feature', hash: 'feature-commit' }]
-          const git = {
-            ...createMockGitClient({ commits: baseCommits }),
-            getCommitLog: (opts?: { maxCount?: number; path?: string }) => {
-              if (opts?.path === 'tools/') {
-                return []
-              }
-              if (opts?.maxCount) {
-                return baseCommits.slice(0, opts.maxCount)
-              }
-              return baseCommits
-            },
-          } as unknown as GitClient
-          const logger = createMockLogger()
-          const context = createMockContext({
-            git,
-            logger,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: {
-                infrastructure: {
-                  paths: ['tools/'],
-                },
-              },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          await step.execute(context)
-
-          expect(logger.debug).toHaveBeenCalledWith(expect.stringContaining('Found 0 commits touching infrastructure paths'))
-        })
-      })
-
-      describe('dependency tracking (Phase 4)', () => {
-        it('skips dependency map when trackDependencyChanges is false', async () => {
-          const git = createMockGitClient({
-            commits: [{ message: 'feat: feature', hash: 'commit1' }],
-          })
-          const logger = createMockLogger()
-          const context = createMockContext({
-            git,
-            logger,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: {
-                trackDependencyChanges: false,
-              },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          await step.execute(context)
-
-          const debugCalls = (logger.debug as jest.Mock).mock.calls.map((call) => call[0])
-          const depLogs = debugCalls.filter((msg: string) => msg.includes('dependencies') || msg.includes('Dependency'))
-          expect(depLogs).toHaveLength(0)
-        })
-
-        it('attempts dependency map build when trackDependencyChanges is true', async () => {
-          const git = createMockGitClient({
-            commits: [{ message: 'feat: feature', hash: 'commit1' }],
-          })
-          const logger = createMockLogger()
-          const context = createMockContext({
-            git,
-            logger,
-            state: { isFirstRelease: true },
-            config: {
-              scopeFiltering: {
-                trackDependencyChanges: true,
-              },
-            },
-          })
-          const step = createAnalyzeCommitsStep()
-
-          await step.execute(context)
-
-          const debugCalls = (logger.debug as jest.Mock).mock.calls.map((call) => call[0])
-          const depLogs = debugCalls.filter(
-            (msg: string) => msg.includes('dependencies') || msg.includes('Dependency') || msg.includes('dependency')
-          )
-          expect(depLogs.length).toBeGreaterThan(0)
         })
       })
     })
